[derived=9e9777] Initial Baseline commit ( reworked )
This commit is a restart of my development work to provide an
absolutely minimal baseline for the project library, complete with
the tools required to re-indent all the important aspects of generated
documentation such as they will be coherent when viewed via git
history.

The objective is for each commit to be "complete", including tests,
specifically, tests known not to fail, in order to provide a useful
series of commits for even the most adventurous git bisectors.

Unfortunately to deliver on this I've needed to reach out a fair bit
and implement my own, hap-hazard CSS and HTML formatting tools, because
for some apparent reason, the internet is entirely devoid of tools that
are suited for this task, either having profanity inducing steps for me
to even install, or having limitations that break the generated content
beyond recognition.

This baseline commit is tested to compile and pass tests on Rust 1.31.1
and should work on any later commit, and going forward, adding new
dependencies will include tooling to determine the oldest versions of
those versions that can be used.